travis: Include the PWD and exit code in travis cd error messages
To make failures like the following more clear. There it's clear that
/home/travis/build/bitcoin/bitcoin/depends/work/build/x86_64-apple-darwin14/
exists, but $PWD/build/bitcoin-x86_64-apple-darwin14 does not, or else
there was some other error (as potentially indicated by the exit code).

Note I would prefer to return the exit code, rather than log it, but the
echo return code would get in the way and working around that would be
obtuse.